Got Trace hiding to work in a primitive way (no unhide yet).
The earlier attempts to hide contours (at the series level) were only
able to hide the marker contours. This version was able to hide the
trace contours themselves (which are the contours inside sections).

This version still has lots of experimental code (to be removed), but it
can't "toggle" the hiding/unhiding of all traces yet.

Also, the section isn't automatically redisplayed in this version. So
the hiding doesn't actually happen until the section is left and then
redisplayed.