Implemented Hide All Traces (Control-E) as a toggle per Section.
This commit allows the hiding and unhiding of all traces on the current
section using the "Control-E" key. However, the traces will not be
hidden immediately because the current code does not force the screen to
update. In order to see the actual hiding (or unhiding) the screen needs
to be forced to redraw itself (page up/page down for example).

Future versions will hopefully trigger a redraw when hiding or unhiding.