Java Version: Removed temporary scale testing code.
This version amounts to a cross-platform Reconstruct viewer. It can open
a series file and scroll through the sections showing images and traces.

The drawing functionality is still working as it did in the original
demonstration, but it's incompatible with the XML coordinate system.
In other words, it draws traces in the wrong locations and doesn't add
them to the XML data yet.