Java: Began converting the "active_stroke" to an "active_contour".
The previous version managed the active drawing trace as "stroke" in the
native coordinate system. This meant that the active drawing trace was
of a different type than the contours that had already been drawn.

In anticipation of supporting other representations (like Bezier curves)
as contours, it makes sense to consolidate the trace representations in
one place. This commit shadows the active_stroke with an active_contour
in preparation for removing the active_stroke.

At this time, the active_contour is drawn in white and a bit to the
right of the active_stroke. That will be changed when the active_stroke
is eventually removed.