Added support for exporting contours from the Series file.
This version exports byte-identical series files for the organelle
slices in the data directory. It still does not export the section
files yet.

Note that the exporting of these contours was done by example with
the hard-coded attributes found in that example. There may be other
attribute values that haven't shown up in the example and are not
yet handled.