Finished preliminary code to write out XML for section files.
This commit can generate XML files for sections based on the XML
structures loaded from the original input files. This is done with
the "Save" menu option.

Note that these XML files are valid XML, but they are not whitespace
identical to the files produced by Reconstruct.exe. Also note that the
current code uses its own tables of names to write the XML. This was
done to ensure that the ordering is preserved. However, this may not be
needed if the parsed XML documents preserve ordering (unlike a typical
"dictionary" implementation).