Java: Draw points and calculate areas in dump.
This version draws the actual points in an XML series (should be an
option). This was implemented to show the actual points generated by
Reconstruct.exe which may be more than expected when using "Closed
Point-by-Point" drawing mode (with the default settings).

This version also calculates the area when dumping traces. The area
hasn't been fully verified yet.