Java: Added Series / Convert to CRLF for Reconstruct.exe compatibility.
After considerable investigation, it appears that Reconstruct.exe does
not handle arbitrary white space properly when reading from the XML. In
particular, it assumes CR+LF line endings (carriage return + line feed).

Reconstruct.exe can partially read files using only the Unix line feed
endings, but it ends up with only one or two sections (and ultimately
only one section). This was very puzzling because there were no error
messages printed. After a lot of  trial and error, the line endings were
found to be the cause of the limited section reading.

This appears to be a problem only for the Windows Reconstruct.exe
program because it uses its own XML parsing. Proper XML parsing should
ignore all white space (carriage returns, line feeds, tabs, spaces...).
The Java version uses Java's built in XML parsing code which should be
handling all white space properly.

As a work-around, this commit adds the following menu option:

   Series / Convert to CRLF

This new menu option will prompt for a series file to convert to CR+LF
format. It does this by looking for all LF characters that are NOT
already preceded by a CR character. When such "lone" LF characters are
found, they are replaced by a CR+LF pair. This logic seems to work
reasonably well, and it has the advantage that converting an existing
file already in CR+LF format will keep that CR+LF format. The conversion
is done to both the series file itself (.ser) as well as any section
files sharing the same name (Reconstruct's default behavior). The
conversion does not affect the currently open series in any way. It's
really more of an external utility bundled into Reconstruct.java for
convenience.

The normal usage of this option would happen after saving a series. Then
the series can be converted to CRLF to be imported into Reconstruct.exe.
There is no reverse conversion needed (at least not at this time) since
Reconstruct.java can parse either format.

Note that this problem may be mitigated (or even hidden) when sharing
files through Git. Git has the ability to manage line endings for each
platform automatically. See discussion at:

   https://help.github.com/articles/dealing-with-line-endings/

When Git is handling line endings, it typically stores the files with LF
endings internally. But when the files are checked out in Windows, the
CR+LF endings are applied. And when files are checked back in (in
Windows), the CR is removed.