Added arrow drawing option to show orientation of traces.
The orientation of traces has been greatly confused due to the conflict
between the Reconstruct user's manual and the actual storage of points
in the XML "Contour" element.

The Reconstruct user manual is very clear in serveral places that traces
drawn in the clockwise direction will enclose areas and generate
positive areas (see Chapter 8 - Traces). However, these points appear to
be stored in reverse order in the XML files.

In order to help resolve these issues, this version of Reconstruct.java
includes a "Show Arrows" option in the "Extras / Line" menu. When
enabled, this option will draw small arrows at each point in a trace.
The first point will be marked with a small square.