Reverse trace directions when reading and writing the XML files.
It turns out that Reconstruct.exe appears to store the points in the
reverse order that they were created. Since the "direction" of the
points in a trace is used to deduce whether it's an outer or inner
boundary, this direction must be consistent with Reconstruct.exe.

This commit continues to use traces stored in the "forward" order
internally, but it reverses them when reading and writing XML files.