Change "cc" to "gcc" in gbm/jpeg-6a/makeijgo - SEE BELOW!!
It appears that the current makefile doesn't fully compile the JPEG code in the MINGW environment.
Here's the process:
  - The top level makefile calls make in the gbm directory with the default "makefile".
  - The gbm makefile calls make in the jpeg-6a directory with the "makeijgo" makefile.
  - The "makeijgo" makefile had defined "CC" as "cc" which is incorrect in the MINGW environment.

This commit changed the "makeijgo" file to use "gcc" rather than "cc". But there is another problem.
The cascading of make files does not appear to be operating correctly, and the jpeg-6a files are NOT
being recompiled when only the top level make file is called.

To properly (and fully) compile the JPEG code, cd to the gbm/jpeg-6a directory and execute:

    mingw32-make -f makeijgo

This will properly build the object files in the JPEG library. Then issue the normal top-level make:

    mingw32-make

That should fully rebuild the executable.